+++ src/main/java/com/tnki/core/memox/repository/MemoItemRepository.java
@@ -5,7 +5,7 @@ import com.tnki.core.memox.model.MemoItem;
 import java.util.List;
 
 public interface MemoItemRepository {
-    List<MemoItem> findAllUnLearned(int userID, int limit);
+    List<MemoItem> findAllUnLearned(long userID, int limit);
     void add(MemoItem item, long userID);
     MemoItem findOne(int memoItemID);
 }
