+++ src/main/java/com/tnki/core/memox/repository/MemoUserSettingRepository.java
@@ -5,6 +5,6 @@ import com.tnki.core.memox.model.UserLearnSetting;
 import java.util.Optional;
 
 public interface MemoUserSettingRepository {
-    Optional<UserLearnSetting> findOne(int userID);
+    Optional<UserLearnSetting> findOne(long userID);
     void save(UserLearnSetting userLearnSetting);
 }
