+++ src/main/java/com/tnki/core/statistics/repository/DailyStatisticsRepository.java
@@ -6,7 +6,7 @@ import java.util.Date;
 import java.util.Optional;
 
 public interface DailyStatisticsRepository {
-    Optional<DailyStatistics> findOne(int userID, Date date);
+    Optional<DailyStatistics> findOne(long userID, Date date);
     void add(DailyStatistics dailyStatistics);
     void update(DailyStatistics dailyStatistics);
 }
